server/requirements.txt
before-after==1.0.1
compress-pickle==2.1.0
fastapi==0.100.0
httpx==0.24.1
lz4==4.3.2
mock==1.0.1
msgpack==1.0.5
numpy==1.25.1
orjson==3.9.5
Pillow==10.0.0
pydantic-numpy==5.0.1
pytest==7.4.0
pytest_asyncio==0.21.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
setuptools==68.0.0
starlette==0.27.0
uvicorn==0.23.1
websockets==11.0.3
